backend/requirements.txt
beanie
motor
langchain
langchain-google-genai
langchain-groq
langchain-community
pypdf
pinecone
langchain-pinecone
langchain_core
fastapi
uvicorn
PyJWT
passlib[bcrypt]
python-dotenv
argon2-cffi\npython-docx
